third_party/WebKit/LayoutTests/imported/wpt/offscreen-canvas/fill-and-stroke-styles/2d.fillStyle.parse.css-color-4-hsl-7-expected.txt

This is a testharness.js-based test.
FAIL OffscreenCanvas test: 2d.fillStyle.parse.css-color-4-hsl-7 assert_equals: Red channel of the pixel at (50, 25) expected 0 but got 255
Harness: the test ran to completion.
